Exception handling Scope
The above example provides two handy functions, get_int_in_range and create_list. The get_int_in_range function expects the user to enter an integer value that falls within the range of values specified by its parameters. It ensures the integers the user provides is in the correct range, but if the user enters a non-integer value, the function will raise an exception.
An uncaught exception produces a stack trace. Python uses an area of the computer's memory knows as the stack to help it control function and method invocations. The stacks parameters returns values, and the point in the code where the program's executions should return when a function completes.