Exception Handling Scope
Situations where thighs could go wrong with a function read_file

A file name may not exist, at least in the location expected by the program. 

A file by the given name exist, but the program (or more precisely, the user running the program) is not authorized to read the
file.

The program encountered bad sector on the disk while reading the file

The data in the file is not of the proper format; for example, the data consists of arbitrary string rather than text convertible to integers.

Some other file error exists.