Create Rational Number Class
We will define our own simple rational number class. A rational number is the ratio of two integers. There is a restriction, however--the number on the bottom of a fraction cannot be zero. The number on the top of a fraction is called the numerator, and the bottom number is known as the denominator.